backend/requirements.txt
fastapi==0.115.12
uvicorn[standard]==0.34.3
anthropic==0.52.0
pydantic-settings==2.9.1
Pillow==11.2.1
python-multipart==0.0.20
httpx==0.28.1
pytest==8.3.5
pytest-asyncio==0.26.0
pytest-mock==3.14.0
ruff==0.11.9
